requirements.txt
fastmcp>=0.1.0
mcp>=0.1.0
requests>=2.31.0
aiohttp>=3.8.0
pyyaml>=6.0
gitpython>=3.1.0